fix(wallet): Move Portfolio Settings to Wallet Settings Menu
