+++ components/brave_wallet/browser/brave_wallet_constants.h
@@ -454,6 +454,7 @@ inline constexpr webui::LocalizedString kLocalizedStrings[] = {
     {"braveWalletAccountsExport", IDS_BRAVE_WALLET_ACCOUNTS_EXPORT},
     {"braveWalletAccountsDeposit", IDS_BRAVE_WALLET_ACCOUNTS_DEPOSIT},
     {"braveWalletAccountsRemove", IDS_BRAVE_WALLET_ACCOUNTS_REMOVE},
+    {"braveWalletAccountSettings", IDS_BRAVE_WALLET_ACCOUNT_SETTINGS},
     {"braveWalletNoAvailableAssets", IDS_BRAVE_WALLET_NO_AVAILABLE_ASSETS},
     {"braveWalletNoAvailableAssetsDescription",
      IDS_BRAVE_WALLET_NO_AVAILABLE_ASSETS_DESCRIPTION},
@@ -935,6 +936,7 @@ inline constexpr webui::LocalizedString kLocalizedStrings[] = {
     {"braveWalletShowSpamNftsTitle", IDS_BRAVE_WALLET_SHOW_SPAM_NFTS_TITLE},
     {"braveWalletShowSpamNftsDescription",
      IDS_BRAVE_WALLET_SHOW_SPAM_NFTS_DESCRIPTION},
+    {"braveWalletPortfolioSettings", IDS_BRAVE_WALLET_PORTFOLIO_SETTINGS},
     {"braveWalletAccountFilterAllAccounts",
      IDS_BRAVE_WALLET_ACCOUNT_FILTER_ALL_ACCOUNTS},
     {"braveWalletTransactionSubmittedTitle",
