+++ components/brave_wallet_ui/components/desktop/card-headers/default-panel-header.tsx
@@ -15,7 +15,7 @@ import { useOnClickOutside } from '../../../common/hooks/useOnClickOutside'
 import { openWalletRouteTab } from '../../../utils/routes-utils'
 
 // Components
-import { DefaultPanelMenu } from '../wallet-menus/default-panel-menu'
+import { WalletSettingsMenu } from '../wallet-menus/wallet_settings_menu'
 import {
   DAppConnectionSettings //
 } from '../../extension/dapp-connection-settings/dapp-connection-settings'
@@ -83,7 +83,7 @@ export const DefaultPanelHeader = (props: Props) => {
           <Button onClick={() => setShowSettingsMenu((prev) => !prev)}>
             <ButtonIcon name='more-horizontal' />
           </Button>
-          {showSettingsMenu && <DefaultPanelMenu />}
+          {showSettingsMenu && <WalletSettingsMenu />}
         </MenuWrapper>
       </LeftRightContainer>
       {showSettingsMenu && <ClickAwayArea />}
