+++ components/brave_wallet_ui/page/screens/shared-screen-components/tab-header/tab-header.tsx
@@ -4,13 +4,9 @@
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
-import { useHistory } from 'react-router-dom'
 
 // Hooks
 import { useOnClickOutside } from '../../../../common/hooks/useOnClickOutside'
-
-// Types
-import { WalletRoutes } from '../../../../constants/types'
 
 // Styled Components
 import {
@@ -25,7 +21,7 @@ import { Row } from '../../send/shared.styles'
 // Components
 import {
   WalletSettingsMenu //
-} from '../../../../components/desktop/wallet-menus/wallet-settings-menu'
+} from '../../../../components/desktop/wallet-menus/wallet_settings_menu'
 
 export interface Props {
   hideHeaderMenu?: boolean
@@ -34,19 +30,11 @@ export interface Props {
 export const TabHeader = (props: Props) => {
   const { hideHeaderMenu } = props
 
-  // Routing
-  const history = useHistory()
-
   // State
   const [showSettings, setShowSettings] = React.useState<boolean>(false)
 
   // Refs
   const settingsModalRef = React.useRef<HTMLDivElement>(null)
-
-  // Methods
-  const onShowBackup = () => {
-    history.push(WalletRoutes.Backup)
-  }
 
   // Hooks
   useOnClickOutside(
@@ -71,7 +59,6 @@ export const TabHeader = (props: Props) => {
           </SettingsButton>
           {showSettings && (
             <WalletSettingsMenu
-              onClickBackup={onShowBackup}
               onClosePopup={() => setShowSettings(false)}
               yPosition={50}
             />
